topList now passes which loc was selected between Search and Home layouts
